refactor(optimize): rename eval modules to drop optimize_ prefix, args to params
- optimize_eval.rs -> eval.rs
- optimize_dense_eval.rs -> dense_eval.rs
- optimize_sparse_eval.rs -> sparse_eval.rs
- optimize_indel.rs -> indel.rs
- args.rs -> params.rs (domain enums, not CLI args)
