+++ packages/treetime/src/commands/timetree/run.rs
@@ -17,7 +17,7 @@ use crate::commands::timetree::output::confidence::{
   write_confidence_intervals,
 };
 use crate::commands::timetree::refinement::run_refinement_iteration;
-use crate::optimize::args::BranchOptMethod;
+use crate::optimize::params::BranchOptMethod;
 use crate::optimize::iteration::{apply_damping, save_branch_lengths};
 use crate::optimize::optimize_unified::{run_optimize_mixed, run_optimize_mixed_inner};
 use crate::representation::partition::timetree::GraphTimetree;
