+++ packages/treetime/src/optimize/__tests__/test_coefficient_extraction_sparse/test_coefficient_extraction_sparse_derivatives.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod tests {
   use crate::gtr::get_gtr::{JC69Params, jc69};
-  use crate::optimize::optimize_dense_eval::evaluate_dense_contribution;
-  use crate::optimize::optimize_sparse_eval::evaluate_sparse_contribution;
+  use crate::optimize::dense_eval::evaluate_dense_contribution;
+  use crate::optimize::sparse_eval::evaluate_sparse_contribution;
   use crate::pretty_assert_ulps_eq;
   use crate::representation::partition::optimize_dense;
   use crate::representation::partition::optimize_sparse::{PartitionContribution, SiteContribution};
