+++ packages/treetime/src/optimize/__tests__/test_convergence/test_convergence_idempotence.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
   use crate::ancestral::marginal::update_marginal;
-  use crate::optimize::args::BranchOptMethod;
+  use crate::optimize::params::BranchOptMethod;
   use crate::optimize::optimize_unified::run_optimize_mixed;
   use crate::pretty_assert_ulps_eq;
   use crate::representation::payload::ancestral::GraphAncestral;
