+++ packages/treetime/src/optimize/__tests__/test_damping.rs
@@ -3,7 +3,7 @@ mod tests {
   use crate::optimize::__tests__::test_convergence::test_convergence_support::tests::{
     TREE_NEWICK, compute_total_lh, setup_partitions, simple_alignment,
   };
-  use crate::optimize::args::BranchOptMethod;
+  use crate::optimize::params::BranchOptMethod;
   use crate::optimize::iteration::{apply_damping, save_branch_lengths};
   use crate::optimize::run_loop::run_optimize_loop;
   use crate::representation::payload::ancestral::GraphAncestral;
