+++ packages/treetime/src/optimize/__tests__/test_eval_zero_branch_mismatch.rs
@@ -3,7 +3,7 @@ mod tests {
   use crate::alphabet::alphabet::{Alphabet, AlphabetName};
   use crate::ancestral::marginal::{initialize_marginal, update_marginal};
   use crate::gtr::get_gtr::{JC69Params, jc69};
-  use crate::optimize::args::BranchOptMethod;
+  use crate::optimize::params::BranchOptMethod;
   use crate::optimize::optimize_unified::run_optimize_mixed;
   use crate::optimize::run_loop::collect_optimize_partitions;
   use crate::representation::partition::fitch::PartitionFitch;
