+++ packages/treetime/src/optimize/__tests__/test_gm_optimize.rs
@@ -6,7 +6,7 @@ mod tests {
   use rstest::rstest;
   use std::collections::BTreeMap;
 
-  use crate::optimize::args::BranchOptMethod;
+  use crate::optimize::params::BranchOptMethod;
   use std::path::Path;
   use treetime_graph::edge::HasBranchLength;
 
@@ -174,7 +174,7 @@ mod tests {
     use crate::alphabet::alphabet::Alphabet;
     use crate::ancestral::marginal::{initialize_marginal, update_marginal};
     use crate::gtr::get_gtr::{JC69Params, jc69};
-    use crate::optimize::args::BranchOptMethod;
+    use crate::optimize::params::BranchOptMethod;
     use crate::optimize::optimize_unified::initial_guess_mixed;
     use crate::optimize::run_loop::{collect_optimize_partitions, run_optimize_loop};
     use crate::representation::partition::fitch::PartitionFitch;
