+++ packages/treetime/src/optimize/__tests__/test_optimize_zero_sequence_length.rs
@@ -2,7 +2,7 @@
 mod tests {
   use crate::alphabet::alphabet::{Alphabet, AlphabetName};
   use crate::gtr::get_gtr::{JC69Params, jc69};
-  use crate::optimize::args::BranchOptMethod;
+  use crate::optimize::params::BranchOptMethod;
   use crate::optimize::optimize_unified::{
     initial_guess_mixed, run_optimize_mixed, run_optimize_mixed_with_indel_rate,
   };
