+++ packages/treetime/src/optimize/iteration.rs
@@ -1,4 +1,4 @@
-use crate::optimize::args::InitialGuessMode;
+use crate::optimize::params::InitialGuessMode;
 use crate::optimize::optimize_unified::initial_guess_mixed;
 use crate::representation::algo::topology_cleanup::collapse::collapse_edge;
 use crate::representation::algo::topology_cleanup::merge_shared_mutations::merge_shared_mutation_branches;
