+++ packages/treetime/src/optimize/mod.rs
@@ -1,13 +1,13 @@
 #[cfg(test)]
 mod __tests__;
 
-pub mod args;
+pub mod dense_eval;
+pub mod eval;
+pub mod indel;
 pub mod iteration;
 pub mod method_brent;
 pub mod method_newton;
-pub mod optimize_dense_eval;
-pub mod optimize_eval;
-pub mod optimize_indel;
-pub mod optimize_sparse_eval;
 pub mod optimize_unified;
+pub mod params;
 pub mod run_loop;
+pub mod sparse_eval;
